+++ src/components/upload/AlbumCover.vue
